Limit the x-axis acceleration of AVs using safe velocities for Potential Lines model
This commit implements safe velocity, in order to avoid the AVs getting too close to the vehicles in the front. The formulation is same as the one used for strip based human model. The closest vehicle whose width in the y-axis overlaps with the ego is considered leader for the model.
The safe velocity is only applied to the model list provided in variable VSafeVehModels. These model names must correspond to the model types provided in section "Vehicle Models" of the config file.