Leave boundary space equivalent vehicle widths when defining the adaptive potential lines
When modifying the potential lines in the vicinity of HDVs, the whole space in between HDVs was used. However, since the CAV align their centers to the PLs, this approach limited the CAVs from over taking the HDVs.
Therefore, this commit leaves some space on both sides of individual region when defining the adaptive PLs. At the moment, the vehicle width of largest vehicle in the simulation, i.e. 1.9m, is used for defining the boundary space.
Any region in between HDVs that has less space than the vehicle width is not used for adaptive PL.