Use road_boundary functions for boundary control and assigning potential lines instead of the fixed parameters LowerLong and UpperLong
Previously, the parameters LowerLong and UpperLong given in the config file were used as road boundaries. This was used to calculate the Potential Lines (PLs) and inside the PD controller for boundary control.
Now, the road boundaries calculated using the functions provided by the TrafficFluidSim. Refer to the manual of TrafficFluidSim. The boundaries for each vehicle flow is defined in the "route" file.
Three pramaters are also introduced if a logitudinal offset should be taken into account while calculating the road boundaries for PLs calculations and the boundary control.